requirements.txt
Flask==3.0.0
opencv-python>=4.12.0
python-pptx==0.6.23
reportlab>=4.0.0
numpy>=2.0.0,<2.3.0
Pillow>=10.2.0
Werkzeug==3.0.1
# TensorFlow is optional - required only for ML-based feature extraction
# If you encounter DLL errors on Windows, comment out or remove the line below
# The application will automatically fall back to SSIM-based scene detection
tensorflow>=2.15.0
scikit-image>=0.21.0

